Updated to version 2.0.0. Compatible with Fork version 3.8.7.
