--- backend/modules/mini_blog/config.php
@@ -1,31 +1,0 @@
-<?php
-
-/*
- * This file is part of Fork CMS.
- *
- * For the full copyright and license information, please view the license
- * file that was distributed with this source code.
- */
-
-/**
- * This action will add a post to the blog module.
- *
- * @author Stef Bastiaansen <stef.bastiaansen@netlash.com>
- */
-final class BackendMiniBlogConfig extends BackendBaseConfig
-{
-	/**
-	 * The default action
-	 *
-	 * @var	string
-	 */
-	protected $defaultAction = 'index';
-
-
-	/**
-	 * The disabled actions
-	 *
-	 * @var	array
-	 */
-	protected $disabledActions = array();
-}
