--- frontend/modules/mini_blog/config.php
@@ -1,31 +1,0 @@
-<?php
-
-/*
- * This file is part of Fork CMS.
- *
- * For the full copyright and license information, please view the license
- * file that was distributed with this source code.
- */
-
-/**
- * This is the configuration-object
- *
- * @author Tijs Verkoyen <tijs@sumocoders.be>
- */
-final class FrontendMiniBlogConfig extends FrontendBaseConfig
-{
-	/**
-	 * The default action
-	 *
-	 * @var	string
-	 */
-	protected $defaultAction = 'index';
-
-
-	/**
-	 * The disabled actions
-	 *
-	 * @var	array
-	 */
-	protected $disabledActions = array();
-}
