+++ src/Backend/Modules/MiniBlog/Installer/Data/install.sql
