+++ src/Frontend/Modules/MiniBlog/Layout/Widgets/RecentPosts.tpl
